refactor: rename component files and update import paths for consistency across next-dedot and next-papi templates
