+++ templates/next-dedot/app/hooks/use-connect.ts
