+++ templates/next-dedot/app/hooks/use-current-block.ts
