+++ templates/next-papi/app/hooks/use-current-block.ts
